requirements.txt
# HTTP Client libraries.
aiohttp
requests

# To get the K8s access token for Google Kubernetes Engine (GKE).
google-api-python-client

# Miscellaneous.
pyyaml
flask
